+++ src/hidden.ts
@@ -1,4 +1,5 @@
 import {hiddenToken} from '../mixin/hidden';
+import {noEscape} from '../mixin/noEscape';
 import {Token} from './index';
 
 /* NOT FOR BROWSER */
@@ -14,7 +15,7 @@ import type {Config} from '../base';
  *
  * 不可见的节点
  */
-@hiddenToken()
+@hiddenToken() @noEscape
 export class HiddenToken extends Token {
 	override get type(): 'hidden' {
 		return 'hidden';